Co-locate injectable service contracts
