+++ extensions/mssql/src/azure/msal/msalAzureController.ts
@@ -29,7 +29,7 @@ import {
 } from "../providerSettings";
 import { Deferred } from "../../protocol";
 import { IPrompter } from "../../prompts/question";
-import { ICredentialStore } from "../../credentialstore/icredentialstore";
+import { ICredentialStore } from "../../credentialstore/credentialstore";
 import * as azureUtils from "../utils";
 import { ILogger } from "../../sharedInterfaces/logger";
 import { sendActionEvent } from "extension-toolkit/vscode";
