+++ extensions/mssql/src/connectionconfig/connectionconfig.ts
@@ -8,8 +8,8 @@ import * as LocalizedConstants from "../constants/locConstants";
 import * as Utils from "../models/utils";
 import * as vscode from "vscode";
 import { uuid } from "../utils/utils";
+import { createServiceIdentifier } from "extension-toolkit/base";
 import { IConnectionGroup, IConnectionProfile } from "../models/interfaces";
-import { IConnectionConfig } from "./iconnectionconfig";
 import { ConnectionProfile } from "../models/connectionProfile";
 import { getConnectionDisplayName } from "../models/connectionInfo";
 import { Deferred } from "../protocol";
@@ -18,6 +18,43 @@ import { logger } from "../models/logger";
 import { ConfigurationTarget } from "vscode";
 
 export type ConfigTarget = ConfigurationTarget.Global | ConfigurationTarget.Workspace;
+
+/**
+ * Interface for a configuration file that stores connection profiles
+ *
+ * @export
+ * @interface IConnectionConfig
+ */
+export const IConnectionConfig = createServiceIdentifier<IConnectionConfig>("connectionConfig");
+
+export interface IConnectionConfig {
+    readonly _serviceBrand: undefined;
+
+    /**
+     * Resolves once the connection config has finished loading connections and groups
+     * from settings and any required migration/initialization has completed.
+     */
+    readonly initialized: Deferred<void>;
+
+    getConnections(): Promise<IConnectionProfile[]>;
+    getConnectionById(id: string): Promise<IConnectionProfile | undefined>;
+    addConnection(profile: IConnectionProfile): Promise<void>;
+    removeConnection(profile: IConnectionProfile): Promise<boolean>;
+    updateConnection(updatedProfile: IConnectionProfile): Promise<void>;
+
+    /**
+     * Populates any missing metadata (group, ID, config source) on the given profile.
+     * @returns true if the profile was modified.
+     */
+    populateMissingConnectionMetadata(profile: IConnectionProfile): boolean;
+
+    getRootGroup(): IConnectionGroup | undefined;
+    getGroups(location?: ConfigTarget): Promise<IConnectionGroup[]>;
+    getGroupById(id: string): IConnectionGroup | undefined;
+    addGroup(group: IConnectionGroup): Promise<void>;
+    removeGroup(id: string, contentAction?: "delete" | "move"): Promise<boolean>;
+    updateGroup(updatedGroup: IConnectionGroup): Promise<void>;
+}
 
 /**
  * Implements connection profile file storage.
