+++ extensions/mssql/src/controllers/mainController.ts
@@ -43,8 +43,7 @@ import { ObjectManagementService } from "../services/objectManagementService";
 import StatusView from "../views/statusView";
 import { IConnectionGroup, IConnectionProfile, ISelectionData } from "../models/interfaces";
 import ConnectionManager from "./connectionManager";
-import { ICredentialStore } from "../credentialstore/icredentialstore";
-import { CredentialStore } from "../credentialstore/credentialstore";
+import { CredentialStore, ICredentialStore } from "../credentialstore/credentialstore";
 import { IConnectionStore, ConnectionStore } from "../models/connectionStore";
 import { IInstantiationService } from "extension-toolkit/base";
 import SqlDocumentService, { ConnectionStrategy } from "./sqlDocumentService";
