+++ extensions/mssql/test/unit/mainController.test.ts
@@ -12,10 +12,8 @@ import { InstantiationServiceBuilder, ServiceDescriptor } from "extension-toolki
 import { ExtensionContextService, IExtensionContextService } from "extension-toolkit/vscode";
 import MainController from "../../src/controllers/mainController";
 import ConnectionManager from "../../src/controllers/connectionManager";
-import { CredentialStore } from "../../src/credentialstore/credentialstore";
-import { ICredentialStore } from "../../src/credentialstore/icredentialstore";
-import { ConnectionConfig } from "../../src/connectionconfig/connectionconfig";
-import { IConnectionConfig } from "../../src/connectionconfig/iconnectionconfig";
+import { CredentialStore, ICredentialStore } from "../../src/credentialstore/credentialstore";
+import { ConnectionConfig, IConnectionConfig } from "../../src/connectionconfig/connectionconfig";
 import { ConnectionStore, IConnectionStore } from "../../src/models/connectionStore";
 import { stubTelemetry, stubExtensionContext, stubMessageBoxes } from "./utils";
 import * as Constants from "../../src/constants/constants";
